refactor: P0 architecture optimization — AI dual-API unification + Plugin class split
## P0-A: AI Dual API Consolidation

- Delete legacy Agent class (1199 lines) and its tests (760 lines)
- Extract userMessageToFilterText() as standalone utility function
- Migrate ChatMessage → AgentMessage in 4 files (prompt.ts, prompt-assembly.ts, session-io.ts, task-continuation.ts)
- Replace AgentResult['toolCalls'] with inline type in subagent.ts
- Remove Agent/createAgent/formatToolTitle/AgentState/AgentEvents exports from ai, agent, core, and zhin packages

Net reduction: ~1,959 lines of legacy code removed with 0 test regressions.

Bridge layer (registerLlmApiFromProviders, convertLegacyTools, agentMessagesToOpenAi) retained — actively used by agentLoop until providers migrate to new registry.

## P0-B: Plugin Class Split (Phase 1)

- Extract AsyncLocalStorage, getCurrentFile(), getPlugin() into plugin-context.ts
- Plugin class imports and re-exports from plugin-context.ts
- All existing import paths preserved (backward compatible)

Plugin class reduced from 815 to ~680 lines.

## Verification

- pnpm build: 49/49 tasks successful
- pnpm test: 343/344 test files pass, 3670+ tests pass (1 flaky test unrelated to changes)
