--- packages/im/ai/src/agent/index.ts
@@ -1,1198 +1,0 @@
-/**
- * @zhin.js/agent - Agent System
- * AI Agent 实现，支持工具调用和多轮对话
- */
-
-import { formatCompact, formatCompactUsage, Logger, truncatePreview } from '@zhin.js/logger';
-import type {
-  AIProvider,
-  AgentConfig,
-  AgentTool,
-  AgentResult,
-  ChatMessage,
-  ContentPart,
-  ToolDefinition as ChatToolDefinition,
-  ToolCall,
-  ToolFilterOptions,
-  Usage,
-} from '../types.js';
-import { filterTools } from './tool-filter.js';
-import { mergeToolsByName } from './tool-naming.js';
-import {
-  autoCompactIfNeeded,
-  createAutoCompactTracking,
-  estimateMessagesTokens,
-  microCompactMessages,
-} from '../compaction/index.js';
-import type { AutoCompactTrackingState } from '../compaction/index.js';
-import {
-  compactMediaToolJsonForModel,
-  isMediaToolWithBinaryPayload,
-  parseMediaToolResultForOutbound,
-  sanitizeToolResult,
-} from './tool-result-sanitizer.js';
-import {
-  DEFAULT_POLICY_DENIAL_STOP_AFTER,
-  isPolicyDenialMessage,
-  isPolicyDenialToolResult,
-  POLICY_DENIAL_TOOL_HINT,
-  SYSTEM_POLICY_DENIAL_STOP,
-} from './policy-denial.js';
-
-const logger = new Logger(null, 'Agent');
-
-const LOG_PREVIEW_CHARS = 480;
-
-function previewLogText(value: unknown, max = LOG_PREVIEW_CHARS): string {
-  if (value == null) return '';
-  const text = typeof value === 'string' ? value : JSON.stringify(value);
-  return truncatePreview(text, max);
-}
-
-function logAgentThinking(iter: number, reasoning: unknown): void {
-  const thinking = previewLogText(reasoning);
-  if (!thinking) return;
-  logger.debug(formatCompact({ iter, thinking }));
-}
-
-function logAgentAnswer(iter: number, content: string): void {
-  const answer = previewLogText(content);
-  if (!answer) return;
-  logger.debug(formatCompact({ iter, answer }));
-}
-
-/** 将 LLM 返回的 assistant 消息写入多轮 history（保留 tool_calls 与 reasoning_content） */
-function buildAssistantHistoryMessage(choiceMessage: ChatMessage): ChatMessage {
-  const content = typeof choiceMessage.content === 'string' ? choiceMessage.content : '';
-  const next: ChatMessage = {
-    role: 'assistant',
-    content,
-    ...(choiceMessage.tool_calls?.length ? { tool_calls: choiceMessage.tool_calls } : {}),
-  };
-  if (choiceMessage.reasoning_content != null) {
-    next.reasoning_content = choiceMessage.reasoning_content;
-  }
-  return next;
-}
-
-/** 工具返回后一轮内提示：一行，避免与 Zhin 长 system 重复堆叠 */
-const SYSTEM_TOOL_FOLLOWUP_ANSWER =
-  'From the tool results above, answer the user in natural language; do not call web_search again unless those results were empty or useless.';
-
-const SYSTEM_TOOL_DUP_FORCE = 'Reply in natural language now; do not call more tools.';
-
-/** 工具执行默认超时时间 (ms) */
-const DEFAULT_TOOL_TIMEOUT = 30_000;
-
-/** 默认最大并发工具执行数 */
-const DEFAULT_MAX_CONCURRENT_TOOLS = 10;
-
-/**
- * 判断工具是否可并发执行
- *
- * 优先使用显式的 isConcurrencySafe 标记；
- * 其次使用 isReadOnly 推断（只读工具默认可并发）；
- * 无标记时默认不可并发（fail-closed）。
- */
-function isToolConcurrencySafe(tool: AgentTool): boolean {
-  if (tool.isConcurrencySafe !== undefined) return tool.isConcurrencySafe;
-  if (tool.isReadOnly !== undefined) return tool.isReadOnly;
-  return false; // fail-closed: 未标记的工具默认独占执行
-}
-
-/**
- * 根据工具名和参数生成简短标题（用于日志、TOOLS.md 等）
- */
-export function formatToolTitle(name: string, args?: Record<string, any>): string {
-  if (!args || Object.keys(args).length === 0) return name;
-  const a = args;
-  switch (name) {
-    case 'bash': return a.command != null ? `bash: ${String(a.command).slice(0, 60)}` : name;
-    case 'read_file': return a.file_path != null ? `read_file: ${a.file_path}` : name;
-    case 'write_file': return a.file_path != null ? `write_file: ${a.file_path}` : name;
-    case 'edit_file': return a.file_path != null ? `edit_file: ${a.file_path}` : name;
-    case 'list_dir': return a.path != null ? `list_dir: ${a.path}` : name;
-    case 'web_search': return a.query != null ? `web_search: ${String(a.query).slice(0, 40)}` : name;
-    case 'web_fetch': return a.url != null ? `web_fetch: ${String(a.url).slice(0, 50)}` : name;
-    default: {
-      const first = Object.values(a)[0];
-      if (first != null) return `${name}: ${String(first).slice(0, 50)}`;
-      return name;
-    }
-  }
-}
-
-/**
- * Agent 执行状态
- */
-export interface AgentState {
-  messages: ChatMessage[];
-  toolCalls: { tool: string; args: Record<string, any>; result: any }[];
-  usage: Usage;
-  iterations: number;
-}
-
-/**
- * Agent 事件
- */
-export interface AgentEvents {
-  'thinking': (message: string) => void;
-  'tool_call': (tool: string, args: Record<string, any>) => void;
-  'tool_result': (tool: string, result: any) => void;
-  'streaming': (content: string) => void;
-  'complete': (result: AgentResult) => void;
-  'error': (error: Error) => void;
-  /** 循环内压缩完成时触发 */
-  'compaction': (info: { microSavedTokens: number; autoSavedTokens: number; totalTokensBefore: number; totalTokensAfter: number }) => void;
-}
-
-/**
- * AI Agent 类
- * 支持工具调用、多轮对话、流式输出
- */
-export class Agent {
-  private provider: AIProvider;
-  private config: Required<Pick<AgentConfig, 'provider' | 'model' | 'systemPrompt' | 'tools' | 'maxIterations' | 'temperature'>> & Pick<AgentConfig, 'contextWindow' | 'maxConcurrentTools' | 'modelFallbacks' | 'turnTimeout' | 'transformToolResult' | 'forceMicroCompactEachTurn' | 'policyDenialStopAfter'>;
-  private tools: Map<string, AgentTool> = new Map();
-  private eventHandlers: Map<keyof AgentEvents, Function[]> = new Map();
-
-  constructor(provider: AIProvider, config: AgentConfig) {
-    this.provider = provider;
-    this.config = {
-      provider: config.provider,
-      model: config.model || provider.models[0],
-      modelFallbacks: config.modelFallbacks,
-      systemPrompt: config.systemPrompt || this.getDefaultSystemPrompt(),
-      tools: config.tools || [],
-      maxIterations: config.maxIterations || 10,
-      temperature: config.temperature ?? 0.7,
-      contextWindow: config.contextWindow,
-      maxConcurrentTools: config.maxConcurrentTools,
-      turnTimeout: config.turnTimeout,
-      transformToolResult: config.transformToolResult,
-      forceMicroCompactEachTurn: config.forceMicroCompactEachTurn,
-      policyDenialStopAfter: config.policyDenialStopAfter ?? DEFAULT_POLICY_DENIAL_STOP_AFTER,
-    };
-
-    // 注册工具
-    const merged = mergeToolsByName(this.config.tools, {
-      reservedNames: config.reservedToolNames,
-      reservedPrefixes: config.reservedToolNamePrefixes,
-    });
-    for (const warning of merged.warnings) {
-      logger.warn(formatCompact( {
-        tool_naming: warning.action,
-        name: warning.name,
-        source: warning.source,
-        action: warning.action,
-        reason: warning.reason,
-        ...(warning.previousSource ? { previous: warning.previousSource } : {}),
-      }));
-    }
-    for (const tool of merged.tools) {
-      this.tools.set(tool.name, tool);
-    }
-  }
-
-  /**
-   * Default system prompt (English). Reply language is determined by user profile or message language.
-   */
-  private getDefaultSystemPrompt(): string {
-    return `You are a helpful assistant that can use tools to complete tasks.
-
-# Principles
- - Understand the user's intent and choose appropriate tools
- - If multiple steps are needed, execute them in order
- - On transient tool failure (timeout, network), you may try one alternative
- - On security/policy denial (policyBlocked, ZHIN_NEEDS_OWNER, execAllowlist), do NOT retry with other tools — explain limits to the user
- - If a task cannot be completed, say so honestly
- - NEVER claim to have performed an action unless you called a tool and got a confirmed result
- - If the user asks you to do something you have no tool for, say so honestly — do NOT fabricate a success response
- - Reply in the language specified in [User profile] (key: language / preferred_language), or in the user's message language if not set`;
-  }
-
-  /**
-   * 带自动降级的 chat 调用：主模型失败时依次尝试 modelFallbacks
-   * 每次 LLM 请求独立应用 turnTimeout，而非所有轮次共享。
-   */
-  private async chatWithFallback(request: Omit<import('../types.js').ChatCompletionRequest, 'model'>): Promise<{ response: import('../types.js').ChatCompletionResponse; usedModel: string }> {
-    const candidates = [this.config.model, ...(this.config.modelFallbacks || [])];
-    const turnTimeout = this.config.turnTimeout;
-    let lastError: Error | undefined;
-    for (let i = 0; i < candidates.length; i++) {
-      const model = candidates[i];
-      try {
-        const chatPromise = this.provider.chat({ ...request, model });
-        const response = turnTimeout
-          ? await Promise.race([
-              chatPromise,
-              new Promise<never>((_, rej) =>
-                setTimeout(() => rej(new Error(`LLM 单轮响应超时 (${turnTimeout}ms)`)), turnTimeout),
-              ),
-            ])
-          : await chatPromise;
-        // 验证响应结构：choices 为空视为异常，触发模型降级
-        if (!response.choices?.length) {
-          throw new Error(`模型 ${model} 返回空响应 (choices 为空)`);
-        }
-        // 成功且切换了模型 → 将当前模型提升为首选（后续轮次直接用）
-        if (i > 0) {
-          logger.info(formatCompact( { fallback: `${candidates[0]}→${model}` }));
-          this.config.model = model;
-        }
-        return { response, usedModel: model };
-      } catch (err) {
-        lastError = err as Error;
-        if (i < candidates.length - 1) {
-          logger.warn(formatCompact( {
-            fallback: model,
-            error: truncatePreview(lastError.message),
-            next: candidates[i + 1],
-          }));
-        }
-      }
-    }
-    throw lastError || new Error('All model candidates failed');
-  }
-
-  /**
-   * 注册事件处理器
-   */
-  on<K extends keyof AgentEvents>(event: K, handler: AgentEvents[K]): () => void {
-    const handlers = this.eventHandlers.get(event) || [];
-    handlers.push(handler);
-    this.eventHandlers.set(event, handlers);
-
-    return () => {
-      const idx = handlers.indexOf(handler);
-      if (idx !== -1) handlers.splice(idx, 1);
-    };
-  }
-
-  /**
-   * 触发事件
-   */
-  private emit<K extends keyof AgentEvents>(event: K, ...args: Parameters<AgentEvents[K]>): void {
-    const handlers = this.eventHandlers.get(event) || [];
-    for (const handler of handlers) {
-      try {
-        (handler as Function)(...args);
-      } catch (e) {
-        logger.error('事件处理器错误:', e);
-      }
-    }
-  }
-
-  /**
-   * 添加工具
-   */
-  addTool(tool: AgentTool): void {
-    this.tools.set(tool.name, tool);
-  }
-
-  /**
-   * 移除工具
-   */
-  removeTool(name: string): void {
-    this.tools.delete(name);
-  }
-
-  /**
-   * 释放事件处理器和工具引用，防止内存泄漏。
-   * Agent 通常是单次使用（创建 → run → 丢弃），run() 结束后自动调用
-   */
-  dispose(): void {
-    for (const handlers of this.eventHandlers.values()) {
-      handlers.length = 0;
-    }
-    this.eventHandlers.clear();
-    this.tools.clear();
-  }
-
-  /**
-   * 获取工具定义（缓存在第一次调用后保持不变）
-   */
-  private applyForcedMicroCompact(messages: ChatMessage[]): ChatMessage[] {
-    if (!this.config.forceMicroCompactEachTurn) return messages;
-    const compacted = microCompactMessages(messages, {
-      keepRecentToolResults: 2,
-      force: true,
-    });
-    if (compacted.didCompact) {
-      logger.debug(formatCompact( {
-        micro_compact: compacted.clearedCount,
-        saved: compacted.savedTokens,
-      }));
-    }
-    return compacted.messages;
-  }
-
-  private getToolDefinitions(): ChatToolDefinition[] {
-    return Array.from(this.tools.values()).map(tool => ({
-      type: 'function',
-      function: {
-        name: tool.name,
-        description: tool.description,
-        parameters: tool.parameters,
-      },
-    }));
-  }
-
-  /**
-   * 生成工具调用的去重 key（规范化参数以避免 "" vs "{}" 等差异）
-   */
-  private static toolCallKey(name: string, args: string): string {
-    let normalized: string;
-    try {
-      const parsed = JSON.parse(args || '{}');
-      normalized = JSON.stringify(parsed, Object.keys(parsed).sort());
-    } catch {
-      normalized = args || '';
-    }
-    return `${name}::${normalized}`;
-  }
-
-  /**
-   * 安全解析 JSON，失败则返回原始字符串
-   */
-  private static safeParse(str: string): any {
-    try {
-      return JSON.parse(str);
-    } catch {
-      return str;
-    }
-  }
-
-  /** 从 multimodal user 消息提取用于 tool filter 的纯文本 */
-  static userMessageToFilterText(message: string | ContentPart[]): string {
-    if (typeof message === 'string') return message;
-    return message
-      .filter((p): p is ContentPart & { type: 'text'; text: string } => p.type === 'text' && Boolean(p.text?.trim()))
-      .map(p => p.text)
-      .join('\n');
-  }
-
-  /**
-   * 程序化工具过滤 —— TF-IDF 加权的相关性评分
-   */
-  static filterTools(
-    message: string,
-    tools: AgentTool[],
-    options?: ToolFilterOptions,
-  ): AgentTool[] {
-    return filterTools(message, tools, options);
-  }
-
-  /**
-   * 执行单个工具调用（带超时保护）
-   */
-  private async executeToolCall(toolCall: ToolCall): Promise<string> {
-    const tool = this.tools.get(toolCall.function.name);
-    if (!tool) {
-      return JSON.stringify({
-        error: `Unknown tool: ${toolCall.function.name}`,
-        hint: '该工具不存在，请尝试使用其他可用工具，或直接回答用户。',
-      });
-    }
-
-    let args: Record<string, unknown>;
-    try {
-      args = JSON.parse(toolCall.function.arguments);
-    } catch {
-      return JSON.stringify({
-        error: 'Invalid tool arguments JSON',
-        tool: toolCall.function.name,
-        hint: '请检查工具参数格式后重试。',
-      });
-    }
-
-    logger.debug({ tool: toolCall.function.name, params: args }, 'Executing tool');
-    this.emit('tool_call', tool.name, args);
-
-    try {
-      // 带超时的工具执行
-      const timeout = tool.timeout ?? DEFAULT_TOOL_TIMEOUT;
-      const result = await Promise.race([
-        tool.execute(args),
-        new Promise<never>((_, reject) =>
-          setTimeout(() => reject(new Error(`工具 ${tool.name} 执行超时`)), timeout),
-        ),
-      ]);
-
-      this.emit('tool_result', tool.name, result);
-      return typeof result === 'string' ? result : JSON.stringify(result);
-    } catch (error) {
-      const errorMsg = error instanceof Error ? error.message : String(error);
-      logger.warn(formatCompact( { tool: toolCall.function.name, error: truncatePreview(errorMsg) }));
-      logger.error({ tool: toolCall.function.name, params: args, err: error }, 'Tool execution failed');
-      const policyBlocked = isPolicyDenialMessage(errorMsg);
-      return JSON.stringify({
-        error: errorMsg,
-        tool: toolCall.function.name,
-        ...(policyBlocked ? { policyBlocked: true } : {}),
-        hint: policyBlocked
-          ? POLICY_DENIAL_TOOL_HINT
-          : '该工具执行失败。若为临时错误可换参数重试一次；若与安全策略相关请勿换工具重试，直接向用户说明。',
-      });
-    }
-  }
-
-  /**
-   * 并行执行多个工具调用（跳过重复的）
-   *
-   * 并发控制策略：
-   * - 所有工具分为"并发安全"和"独占"两类
-   * - 并发安全工具可同时执行（上限 maxConcurrentTools，默认 10）
-   * - 独占工具串行执行，等待前面所有工具完成后再开始
-   *
-   * @returns 新执行的工具调用结果列表；如果全部重复则返回空数组
-   */
-  private async executeToolCalls(
-    toolCalls: ToolCall[],
-    seenKeys: Set<string>,
-    state: AgentState,
-  ): Promise<{ toolCall: ToolCall; result: string; args: Record<string, any> }[]> {
-    // 分离：新调用 vs 重复调用
-    const fresh: ToolCall[] = [];
-    for (const tc of toolCalls) {
-      const key = Agent.toolCallKey(tc.function.name, tc.function.arguments);
-      if (seenKeys.has(key)) {
-        logger.debug(`跳过重复工具调用: ${tc.function.name}`);
-      } else {
-        fresh.push(tc);
-      }
-    }
-
-    if (fresh.length === 0) return [];
-
-    const maxConcurrent = this.config.maxConcurrentTools ?? DEFAULT_MAX_CONCURRENT_TOOLS;
-
-    // 将工具分为并发组和独占组
-    const concurrentCalls: ToolCall[] = [];
-    const exclusiveCalls: ToolCall[] = [];
-
-    for (const tc of fresh) {
-      const tool = this.tools.get(tc.function.name);
-      if (tool && isToolConcurrencySafe(tool)) {
-        concurrentCalls.push(tc);
-      } else {
-        exclusiveCalls.push(tc);
-      }
-    }
-
-    const allResults: { toolCall: ToolCall; result: string; args: Record<string, any> }[] = [];
-
-    // 执行单个工具并记录状态
-    const execOne = async (tc: ToolCall) => {
-      let result = await this.executeToolCall(tc);
-      const transform = this.config.transformToolResult;
-      if (transform) {
-        const args = Agent.safeParse(tc.function.arguments);
-        const argsObj = typeof args === 'object' && args != null && !Array.isArray(args)
-          ? (args as Record<string, unknown>)
-          : {};
-        result = await transform({
-          toolName: tc.function.name,
-          toolCallId: tc.id,
-          args: argsObj,
-          result,
-        });
-      }
-      const toolName = tc.function.name;
-      const rawForOutbound = result;
-      const forModel = isMediaToolWithBinaryPayload(toolName)
-        ? compactMediaToolJsonForModel(toolName, rawForOutbound)
-        : rawForOutbound;
-      result = sanitizeToolResult(forModel);
-      const args = Agent.safeParse(tc.function.arguments);
-      const parsedResult = isMediaToolWithBinaryPayload(toolName)
-        ? parseMediaToolResultForOutbound(toolName, rawForOutbound)
-        : Agent.safeParse(result);
-
-      const key = Agent.toolCallKey(toolName, tc.function.arguments);
-      seenKeys.add(key);
-      state.toolCalls.push({
-        tool: toolName,
-        args: typeof args === 'object' ? args : { raw: args },
-        result: parsedResult,
-      });
-
-      return { toolCall: tc, result, args };
-    };
-
-    // ── 并发安全工具：分批并行，每批最多 maxConcurrent ──
-    for (let i = 0; i < concurrentCalls.length; i += maxConcurrent) {
-      const batch = concurrentCalls.slice(i, i + maxConcurrent);
-      const batchResults = await Promise.all(batch.map(execOne));
-      allResults.push(...batchResults);
-    }
-
-    // ── 独占工具：串行执行 ──
-    for (const tc of exclusiveCalls) {
-      const result = await execOne(tc);
-      allResults.push(result);
-    }
-
-    return allResults;
-  }
-
-  /**
-   * 累加 token 用量
-   */
-  private static addUsage(target: Usage, source?: Usage): void {
-    if (!source) return;
-    target.prompt_tokens += source.prompt_tokens;
-    target.completion_tokens += source.completion_tokens;
-    target.total_tokens += source.total_tokens;
-  }
-
-  /** 任务结束时输出 token 汇总与总耗时（debug，主汇总见 AI Handler） */
-  private logTaskComplete(
-    startedAt: number,
-    result: Pick<AgentResult, 'usage' | 'iterations' | 'toolCalls'>,
-    opts?: { model?: string; stream?: boolean },
-  ): void {
-    const elapsedMs = Date.now() - startedAt;
-    const u = result.usage;
-    const toolCount = result.toolCalls?.length ?? 0;
-    logger.debug(formatCompact({
-      total_ms: elapsedMs,
-      usage: formatCompactUsage(u),
-      iter: result.iterations,
-      ...(opts?.stream ? { stream: true } : {}),
-      ...(toolCount > 0 ? { tools: toolCount } : {}),
-      ...(opts?.model ? { model: opts.model } : {}),
-    }));
-  }
-
-  /**
-   * 运行 Agent
-   *
-   * @param userMessage    用户消息
-   * @param context        对话上下文
-   * @param filterOptions  工具过滤选项 —— 启用后在 AI 调用之前程序化筛选工具，省去额外的 AI 意图分析往返
-   */
-  async run(
-    userMessage: string | ContentPart[],
-    context?: ChatMessage[],
-    filterOptions?: ToolFilterOptions,
-  ): Promise<AgentResult> {
-    const startedAt = Date.now();
-    let lastUsedModel = this.config.model;
-    const state: AgentState = {
-      messages: [
-        { role: 'system', content: this.config.systemPrompt },
-        ...(context || []),
-        { role: 'user', content: userMessage },
-      ],
-      toolCalls: [],
-      usage: { prompt_tokens: 0, completion_tokens: 0, total_tokens: 0 },
-      iterations: 0,
-    };
-
-    // 程序化工具预过滤：只把相关工具传给 AI，减少 token 消耗和误选
-    let toolDefinitions: ChatToolDefinition[];
-    if (filterOptions) {
-      const allTools = Array.from(this.tools.values());
-      const filtered = Agent.filterTools(
-        Agent.userMessageToFilterText(userMessage),
-        allTools,
-        filterOptions,
-      );
-      logger.debug(formatCompact( { tools: `${allTools.length}→${filtered.length}` }));
-      toolDefinitions = filtered.map(tool => ({
-        type: 'function' as const,
-        function: {
-          name: tool.name,
-          description: tool.description,
-          parameters: tool.parameters,
-        },
-      }));
-    } else {
-      toolDefinitions = this.getToolDefinitions();
-    }
-    const hasTools = toolDefinitions.length > 0;
-    // O(1) 去重集合
-    const seenToolKeys = new Set<string>();
-    // 连续全重复计数器
-    let consecutiveDuplicateRounds = 0;
-    let policyDenialCount = 0;
-    const policyDenialStopAfter = this.config.policyDenialStopAfter ?? DEFAULT_POLICY_DENIAL_STOP_AFTER;
-
-    // ── 每轮压缩追踪 ──
-    const contextWindow = this.config.contextWindow;
-    const compactTracking = contextWindow ? createAutoCompactTracking() : undefined;
-    let totalMicroSaved = 0;
-    let totalAutoSaved = 0;
-    let compactCount = 0;
-
-    while (state.iterations < this.config.maxIterations) {
-      state.iterations++;
-
-      // ── 每轮入口：上下文压缩 ──
-      // 在调用 LLM 之前检查 token 预算，按需压缩
-      if (contextWindow && compactTracking) {
-        try {
-          const compactResult = await autoCompactIfNeeded({
-            provider: this.provider,
-            messages: state.messages,
-            contextWindow,
-            tracking: compactTracking,
-          });
-          if (compactResult.wasCompacted) {
-            state.messages = compactResult.messages;
-            totalMicroSaved += compactResult.microSavedTokens;
-            totalAutoSaved += compactResult.autoSavedTokens;
-            compactCount++;
-            logger.debug(formatCompact( {
-              iter: state.iterations,
-              compact: compactResult.savedTokens,
-              micro: compactResult.microSavedTokens,
-              auto: compactResult.autoSavedTokens,
-            }));
-            this.emit('compaction', {
-              microSavedTokens: compactResult.microSavedTokens,
-              autoSavedTokens: compactResult.autoSavedTokens,
-              totalTokensBefore: compactResult.savedTokens + estimateMessagesTokens(state.messages),
-              totalTokensAfter: estimateMessagesTokens(state.messages),
-            });
-          }
-        } catch (e) {
-          logger.warn(formatCompact( {
-            iter: state.iterations,
-            compact: 'fail',
-            error: truncatePreview(String(e)),
-          }));
-        }
-      }
-
-      // 强制文本回答的条件：
-      // 1. 检测到连续重复工具调用
-      // 2. 最后一轮迭代且已有工具结果 —— 保证 Agent 始终输出文本，不再需要额外的 summary 往返
-      const isLastIteration = state.iterations >= this.config.maxIterations;
-      const policyCircuitOpen = policyDenialStopAfter > 0 && policyDenialCount >= policyDenialStopAfter;
-      const forceAnswer = consecutiveDuplicateRounds > 0 ||
-        policyCircuitOpen ||
-        (isLastIteration && state.toolCalls.length > 0);
-
-      try {
-        // 工具调用轮次禁用思考
-        const isToolCallRound = hasTools && !forceAnswer;
-        const chatRequest = {
-          messages: state.messages,
-          tools: isToolCallRound ? toolDefinitions : undefined,
-          tool_choice: isToolCallRound ? 'auto' as const : undefined,
-          temperature: this.config.temperature,
-          think: isToolCallRound ? false : undefined,
-        };
-        const { response, usedModel } = await this.chatWithFallback(chatRequest);
-        lastUsedModel = usedModel;
-
-        Agent.addUsage(state.usage, response.usage);
-        logger.debug(formatCompact( {
-          iter: state.iterations,
-          usage: formatCompactUsage(state.usage),
-          model: usedModel,
-        }));
-        const choice = response.choices[0];
-        logAgentThinking(state.iterations, choice.message.reasoning_content);
-
-        // ── 分支 1: 模型想调用工具 ──
-        if (choice.message.tool_calls?.length) {
-          const callSummary = choice.message.tool_calls.map(
-            (tc: any) => `${tc.function.name}(${tc.function.arguments})`
-          ).join(', ');
-          logger.debug(formatCompact( {
-            iter: state.iterations,
-            tool_call: truncatePreview(callSummary),
-          }));
-          this.emit('thinking', '正在执行工具调用...');
-          state.messages.push(buildAssistantHistoryMessage(choice.message));
-
-          // 并行执行所有新工具调用，自动跳过重复
-          const results = await this.executeToolCalls(
-            choice.message.tool_calls,
-            seenToolKeys,
-            state,
-          );
-
-          if (results.length === 0) {
-            consecutiveDuplicateRounds++;
-            logger.warn(formatCompact( { iter: state.iterations, warn: 'dup_tool_calls' }));
-
-            for (const tc of choice.message.tool_calls) {
-              const key = Agent.toolCallKey(tc.function.name, tc.function.arguments);
-              const previous = state.toolCalls.find(
-                stc => Agent.toolCallKey(stc.tool, JSON.stringify(stc.args)) === key ||
-                       Agent.toolCallKey(stc.tool, tc.function.arguments) === key,
-              );
-              state.messages.push({
-                role: 'tool',
-                content: previous ? JSON.stringify(previous.result) : 'Result received.',
-                tool_call_id: tc.id,
-              });
-            }
-
-            state.messages.push({
-              role: 'system',
-              content: SYSTEM_TOOL_DUP_FORCE,
-            });
-
-            continue;
-          }
-
-          // 有新的工具调用被执行
-          consecutiveDuplicateRounds = 0;
-
-          let batchPolicyDenials = 0;
-          // 将工具结果加入消息历史
-          for (const { toolCall, result } of results) {
-            if (isPolicyDenialToolResult(result)) batchPolicyDenials++;
-            logger.debug(formatCompact( {
-              iter: state.iterations,
-              tool_result: toolCall.function.name,
-              preview: previewLogText(result),
-            }));
-            state.messages.push({
-              role: 'tool',
-              content: result,
-              tool_call_id: toolCall.id,
-            });
-          }
-
-          if (batchPolicyDenials > 0) {
-            policyDenialCount += batchPolicyDenials;
-            logger.warn(formatCompact( {
-              iter: state.iterations,
-              warn: 'policy_denial',
-              count: policyDenialCount,
-              stop_after: policyDenialStopAfter,
-            }));
-          }
-
-          if (
-            policyDenialStopAfter > 0 &&
-            policyDenialCount >= policyDenialStopAfter
-          ) {
-            state.messages.push({
-              role: 'system',
-              content: SYSTEM_POLICY_DENIAL_STOP,
-            });
-          } else {
-            // 如果工具返回的是最终结果（非查询中间步骤），引导模型直接回复
-            const allSucceeded = results.every(r => !r.result.startsWith('{'));
-            if (allSucceeded && results.length > 0 && batchPolicyDenials === 0) {
-              state.messages.push({
-                role: 'system',
-                content: SYSTEM_TOOL_FOLLOWUP_ANSWER,
-              });
-            }
-          }
-
-          state.messages = this.applyForcedMicroCompact(state.messages);
-          continue;
-        }
-
-        // ── 分支 2: 模型返回文本回答 ──
-        const content = typeof choice.message.content === 'string'
-          ? choice.message.content
-          : '';
-        logAgentAnswer(state.iterations, content);
-
-        const result: AgentResult = {
-          content,
-          toolCalls: state.toolCalls,
-          usage: state.usage,
-          iterations: state.iterations,
-          model: lastUsedModel,
-          ...(compactCount > 0 && { compaction: { microSavedTokens: totalMicroSaved, autoSavedTokens: totalAutoSaved, compactCount } }),
-        };
-
-        this.logTaskComplete(startedAt, result, { model: lastUsedModel });
-        this.emit('complete', result);
-        return result;
-
-      } catch (error) {
-        const err = error instanceof Error ? error : new Error(String(error));
-        this.emit('error', err);
-
-        // ── 错误恢复策略 ──
-        // 如果已经有工具结果，注入恢复消息让 AI 基于已有数据回答
-        if (state.toolCalls.length > 0) {
-          logger.warn(formatCompact( {
-            iter: state.iterations,
-            recover: true,
-            error: truncatePreview(err.message),
-          }));
-          const toolSummary = state.toolCalls.map(tc => {
-            const r = typeof tc.result === 'string' ? tc.result : JSON.stringify(tc.result);
-            return `【${tc.tool}】${r}`;
-          }).join('\n');
-          const fallbackResult: AgentResult = {
-            content: `${toolSummary}`,
-            toolCalls: state.toolCalls,
-            usage: state.usage,
-            iterations: state.iterations,
-            model: lastUsedModel,
-          };
-          this.logTaskComplete(startedAt, fallbackResult, { model: lastUsedModel });
-          this.emit('complete', fallbackResult);
-          return fallbackResult;
-        }
-
-        // 没有任何工具结果，提供友好的错误消息
-        const fallbackResult: AgentResult = {
-          content: `Something went wrong: ${err instanceof Error ? err.message : String(err)}. Please try again or rephrase your request.`,
-          toolCalls: [],
-          usage: state.usage,
-          iterations: state.iterations,
-          model: lastUsedModel,
-        };
-        this.logTaskComplete(startedAt, fallbackResult, { model: lastUsedModel });
-        this.emit('complete', fallbackResult);
-        return fallbackResult;
-      }
-    }
-
-    // 达到最大迭代次数，基于已有工具结果生成兜底回复
-    let fallbackContent: string;
-    if (state.toolCalls.length > 0) {
-      // 尝试从工具结果中构建有意义的回复
-      const toolSummary = state.toolCalls.map(tc => {
-        const r = typeof tc.result === 'string' ? tc.result : JSON.stringify(tc.result);
-        return `【${tc.tool}】${r}`;
-      }).join('\n');
-      fallbackContent = `Done. Information retrieved:\n${toolSummary}`;
-    } else {
-      fallbackContent = 'Max iterations reached; the task may be incomplete. Try simplifying your request and retry.';
-    }
-
-    const result: AgentResult = {
-      content: fallbackContent,
-      toolCalls: state.toolCalls,
-      usage: state.usage,
-      iterations: state.iterations,
-      model: lastUsedModel,
-      ...(compactCount > 0 && { compaction: { microSavedTokens: totalMicroSaved, autoSavedTokens: totalAutoSaved, compactCount } }),
-    };
-
-    this.logTaskComplete(startedAt, result, { model: lastUsedModel });
-    this.emit('complete', result);
-    return result;
-  }
-
-  /**
-   * 流式运行 Agent
-   *
-   * @param userMessage    用户消息
-   * @param context        对话上下文
-   * @param filterOptions  工具过滤选项 —— 启用后在 AI 调用之前程序化筛选工具
-   */
-  async *runStream(
-    userMessage: string | ContentPart[],
-    context?: ChatMessage[],
-    filterOptions?: ToolFilterOptions,
-  ): AsyncIterable<{
-    type: 'content' | 'tool_call' | 'tool_result' | 'compaction' | 'done';
-    data: any;
-  }> {
-    const startedAt = Date.now();
-    const messages: ChatMessage[] = [
-      { role: 'system', content: this.config.systemPrompt },
-      ...(context || []),
-      { role: 'user', content: userMessage },
-    ];
-    const toolCallHistory: { tool: string; args: any; result: any }[] = [];
-    const seenToolKeys = new Set<string>();
-
-    try {
-
-    // 程序化工具预过滤
-    let toolDefinitions: ChatToolDefinition[];
-    if (filterOptions) {
-      const allTools = Array.from(this.tools.values());
-      const filtered = Agent.filterTools(
-        Agent.userMessageToFilterText(userMessage),
-        allTools,
-        filterOptions,
-      );
-      logger.debug(formatCompact( { tools: `${allTools.length}→${filtered.length}` }));
-      toolDefinitions = filtered.map(tool => ({
-        type: 'function' as const,
-        function: {
-          name: tool.name,
-          description: tool.description,
-          parameters: tool.parameters,
-        },
-      }));
-    } else {
-      toolDefinitions = this.getToolDefinitions();
-    }
-    const hasTools = toolDefinitions.length > 0;
-    let iterations = 0;
-    const usage: Usage = { prompt_tokens: 0, completion_tokens: 0, total_tokens: 0 };
-    let consecutiveDuplicateRounds = 0;
-    let policyDenialCount = 0;
-    const policyDenialStopAfter = this.config.policyDenialStopAfter ?? DEFAULT_POLICY_DENIAL_STOP_AFTER;
-
-    // ── 每轮压缩追踪 ──
-    const contextWindow = this.config.contextWindow;
-    const compactTracking = contextWindow ? createAutoCompactTracking() : undefined;
-    let totalMicroSaved = 0;
-    let totalAutoSaved = 0;
-    let compactCount = 0;
-
-    // 构建与 run() 中 executeToolCalls 兼容的轻量 state
-    const state: AgentState = {
-      messages,
-      toolCalls: toolCallHistory,
-      usage,
-      iterations: 0,
-    };
-
-    while (iterations < this.config.maxIterations) {
-      iterations++;
-      state.iterations = iterations;
-
-      // ── 每轮入口：上下文压缩 ──
-      if (contextWindow && compactTracking) {
-        try {
-          const compactResult = await autoCompactIfNeeded({
-            provider: this.provider,
-            messages,
-            contextWindow,
-            tracking: compactTracking,
-          });
-          if (compactResult.wasCompacted) {
-            // messages 是引用类型，需要原地替换
-            messages.length = 0;
-            messages.push(...compactResult.messages);
-            totalMicroSaved += compactResult.microSavedTokens;
-            totalAutoSaved += compactResult.autoSavedTokens;
-            compactCount++;
-            logger.debug(formatCompact( {
-              iter: iterations,
-              compact: compactResult.savedTokens,
-            }));
-            const info = {
-              microSavedTokens: compactResult.microSavedTokens,
-              autoSavedTokens: compactResult.autoSavedTokens,
-              totalTokensBefore: compactResult.savedTokens + estimateMessagesTokens(messages),
-              totalTokensAfter: estimateMessagesTokens(messages),
-            };
-            this.emit('compaction', info);
-            yield { type: 'compaction', data: info };
-          }
-        } catch (e) {
-          logger.warn(formatCompact( {
-            iter: iterations,
-            compact: 'fail',
-            error: truncatePreview(String(e)),
-          }));
-        }
-      }
-
-      let content = '';
-      let reasoningContent = '';
-      const pendingToolCalls: ToolCall[] = [];
-      let finishReason: string | null = null;
-      const isLastIteration = iterations >= this.config.maxIterations;
-      const policyCircuitOpen = policyDenialStopAfter > 0 && policyDenialCount >= policyDenialStopAfter;
-      const forceAnswer = consecutiveDuplicateRounds > 0 ||
-        policyCircuitOpen ||
-        (isLastIteration && toolCallHistory.length > 0);
-
-      // 流式获取响应
-      for await (const chunk of this.provider.chatStream({
-        model: this.config.model,
-        messages,
-        tools: hasTools && !forceAnswer ? toolDefinitions : undefined,
-        tool_choice: hasTools && !forceAnswer ? 'auto' : undefined,
-        temperature: this.config.temperature,
-      })) {
-        const choice = chunk.choices[0];
-        if (!choice) continue;
-
-        // 处理内容片段
-        if (choice.delta.content) {
-          content += choice.delta.content;
-          // 仅在非工具调用阶段输出内容给消费者
-          if (pendingToolCalls.length === 0) {
-            yield { type: 'content', data: choice.delta.content };
-          }
-        }
-
-        if (choice.delta.reasoning_content) {
-          reasoningContent += choice.delta.reasoning_content;
-        }
-
-        // 合并工具调用片段
-        if (choice.delta.tool_calls) {
-          for (const tc of choice.delta.tool_calls) {
-            let existing = pendingToolCalls.find(p => p.id === tc.id);
-            if (!existing && tc.id) {
-              existing = {
-                id: tc.id,
-                type: 'function',
-                function: { name: '', arguments: '' },
-              };
-              pendingToolCalls.push(existing);
-            }
-            if (existing && tc.function) {
-              if (tc.function.name) existing.function.name += tc.function.name;
-              if (tc.function.arguments) existing.function.arguments += tc.function.arguments;
-            }
-          }
-        }
-
-        if (choice.finish_reason) finishReason = choice.finish_reason;
-
-        Agent.addUsage(usage, chunk.usage);
-      }
-
-      // 将 assistant 消息加入上下文
-      const assistantMsg: ChatMessage = {
-        role: 'assistant',
-        content,
-        tool_calls: pendingToolCalls.length > 0 ? pendingToolCalls : undefined,
-      };
-      if (reasoningContent) assistantMsg.reasoning_content = reasoningContent;
-      messages.push(assistantMsg);
-      logAgentThinking(iterations, reasoningContent);
-
-      // 处理工具调用
-      if (pendingToolCalls.length > 0 && finishReason === 'tool_calls') {
-        // 先通知上层所有工具调用开始
-        for (const tc of pendingToolCalls) {
-          const key = Agent.toolCallKey(tc.function.name, tc.function.arguments);
-          if (!seenToolKeys.has(key)) {
-            yield { type: 'tool_call', data: { name: tc.function.name, args: tc.function.arguments } };
-          }
-        }
-
-        // 使用统一的 executeToolCalls（含读写并发控制 + 去重）
-        const results = await this.executeToolCalls(pendingToolCalls, seenToolKeys, state);
-
-        if (results.length === 0) {
-          // 全部重复
-          consecutiveDuplicateRounds++;
-
-          for (const tc of pendingToolCalls) {
-            const key = Agent.toolCallKey(tc.function.name, tc.function.arguments);
-            const previous = toolCallHistory.find(
-              h => Agent.toolCallKey(h.tool, JSON.stringify(h.args)) === key ||
-                   Agent.toolCallKey(h.tool, tc.function.arguments) === key,
-            );
-            messages.push({
-              role: 'tool',
-              content: previous ? JSON.stringify(previous.result) : 'Result received.',
-              tool_call_id: tc.id,
-            });
-          }
-          continue;
-        }
-
-        consecutiveDuplicateRounds = 0;
-
-        let batchPolicyDenials = 0;
-        // yield 工具结果并加入消息历史
-        for (const { toolCall, result } of results) {
-          if (isPolicyDenialToolResult(result)) batchPolicyDenials++;
-          logger.debug(formatCompact( {
-            iter: iterations,
-            tool_result: toolCall.function.name,
-            preview: previewLogText(result),
-          }));
-          yield { type: 'tool_result', data: { name: toolCall.function.name, result } };
-          messages.push({
-            role: 'tool',
-            content: result,
-            tool_call_id: toolCall.id,
-          });
-        }
-
-        if (batchPolicyDenials > 0) {
-          policyDenialCount += batchPolicyDenials;
-          logger.warn(formatCompact( {
-            iter: iterations,
-            warn: 'policy_denial',
-            count: policyDenialCount,
-            stop_after: policyDenialStopAfter,
-          }));
-        }
-
-        if (
-          policyDenialStopAfter > 0 &&
-          policyDenialCount >= policyDenialStopAfter
-        ) {
-          messages.push({
-            role: 'system',
-            content: SYSTEM_POLICY_DENIAL_STOP,
-          });
-        }
-
-        // 为重复的调用也补上 tool 消息（已被 executeToolCalls 过滤掉的）
-        for (const tc of pendingToolCalls) {
-          const key = Agent.toolCallKey(tc.function.name, tc.function.arguments);
-          const wasExecuted = results.some(r => r.toolCall.id === tc.id);
-          if (!wasExecuted) {
-            const previous = toolCallHistory.find(
-              h => Agent.toolCallKey(h.tool, JSON.stringify(h.args)) === key ||
-                   Agent.toolCallKey(h.tool, tc.function.arguments) === key,
-            );
-            messages.push({
-              role: 'tool',
-              content: previous ? JSON.stringify(previous.result) : 'Result received.',
-              tool_call_id: tc.id,
-            });
-          }
-        }
-
-        const compacted = this.applyForcedMicroCompact(messages);
-        messages.splice(0, messages.length, ...compacted);
-        continue;
-      }
-
-      // 完成
-      logAgentAnswer(iterations, content);
-      const doneData = {
-        content,
-        toolCalls: toolCallHistory,
-        usage,
-        iterations,
-        ...(compactCount > 0 && { compaction: { microSavedTokens: totalMicroSaved, autoSavedTokens: totalAutoSaved, compactCount } }),
-      };
-      this.logTaskComplete(startedAt, doneData, { model: this.config.model, stream: true });
-      yield { type: 'done', data: doneData };
-      return;
-    }
-
-    // 达到最大迭代次数
-    const maxIterData = {
-      content: toolCallHistory.length > 0
-        ? `Done. Executed ${toolCallHistory.length} tool call(s).`
-        : 'Max iterations reached.',
-      toolCalls: toolCallHistory,
-      usage,
-      iterations,
-      ...(compactCount > 0 && { compaction: { microSavedTokens: totalMicroSaved, autoSavedTokens: totalAutoSaved, compactCount } }),
-    };
-    this.logTaskComplete(startedAt, maxIterData, { model: this.config.model, stream: true });
-    yield { type: 'done', data: maxIterData };
-    } finally {
-      messages.length = 0;
-      seenToolKeys.clear();
-      toolCallHistory.length = 0;
-    }
-  }
-}
-
-/**
- * 创建 Agent 实例
- */
-export function createAgent(provider: AIProvider, config: Omit<AgentConfig, 'provider'>): Agent {
-  return new Agent(provider, {
-    ...config,
-    provider: provider.name,
-  });
-}
